Added support for phase filtering of the family accounts count query (country tab) + refactoring
