+++ server/stats/helpers/phaseWasSelected.helper.ts
@@ -1,0 +1,5 @@
+import { VillagePhase } from '../../entities/village';
+
+export const phaseWasSelected = (phase: number | undefined): boolean => {
+  return phase !== undefined && Object.values(VillagePhase).includes(+phase);
+};
